requirements.txt
aiofile==3.8.1
aiofiles==22.1.0
alembic==1.8.1
anyio==3.6.1
appdirs==1.4.4
arq @ git+https://github.com/samuelcolvin/arq@1495be6c234509df1dfdc0db620388d909c8cdb8
async-timeout==4.0.2
async-unzip==0.2.1
asyncpg==0.26.0
caio==0.9.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
fastcrc==0.2.1
gino==1.0.1
h11==0.12.0
hiredis==2.0.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
humanize==4.4.0
idna==3.4
ijson==3.1.4
latest-user-agents==0.0.3
Mako==1.2.3
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
packaging==21.3
pyaml==21.10.1
pylightxl==1.60
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-socks==2.0.3
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
sanic==22.6.2
sanic-routing==22.3.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==1.3.24
typing_extensions==4.3.0
ujson==5.5.0
urllib3==1.26.12
uvloop==0.17.0
websockets==10.3
wrapt==1.14.1
